[PackageBuilder] Move YamlFileLoader to FileLoader namespace, unite Parameters menaming singular
