+++ packages/EasyCodingStandard/tests/Yaml/FileLoader/CheckerTolerantYamlFileLoader/DefinitionsSource/checkers-parent-2.yml
