+++ packages/EasyCodingStandard/tests/Yaml/FileLoader/CheckerTolerantYamlFileLoader/DefinitionsSource/checkers-parent.yml
