+++ packages/EasyCodingStandard/tests/Yaml/FileLoader/CheckerTolerantYamlFileLoader/DefinitionsSource/checkers.yml
