+++ packages/EasyCodingStandard/tests/Yaml/FileLoader/CheckerTolerantYamlFileLoader/DefinitionsSource/config-classic.yml
