+++ packages/EasyCodingStandard/tests/Yaml/FileLoader/CheckerTolerantYamlFileLoader/DefinitionsSource/config-with-at.yml
