+++ packages/PackageBuilder/src/Yaml/FileLoader/ParameterImportsYamlFileLoader.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace Symplify\PackageBuilder\Yaml\FileLoader;
+
+final class ParameterImportsYamlFileLoader extends AbstractParameterImportsYamlFileLoader
+{
+}
