+++ packages/PackageBuilder/src/Yaml/FileLoader/ParameterMergingYamlFileLoader.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Symplify\PackageBuilder\Yaml;
+namespace Symplify\PackageBuilder\Yaml\FileLoader;
 
 final class ParameterMergingYamlFileLoader extends AbstractParameterMergingYamlFileLoader
 {
