+++ packages/Statie/src/DependencyInjection/StatieKernel.php
@@ -16,7 +16,7 @@ use Symplify\PackageBuilder\DependencyInjection\CompilerPass\AutowireDefaultComp
 use Symplify\PackageBuilder\DependencyInjection\CompilerPass\AutowireSinglyImplementedCompilerPass;
 use Symplify\PackageBuilder\DependencyInjection\CompilerPass\PublicDefaultCompilerPass;
 use Symplify\PackageBuilder\DependencyInjection\CompilerPass\PublicForTestsCompilerPass;
-use Symplify\PackageBuilder\Yaml\ParameterMergingYamlFileLoader;
+use Symplify\PackageBuilder\Yaml\FileLoader\ParameterMergingYamlFileLoader;
 use Symplify\Statie\DependencyInjection\CompilerPass\CollectorCompilerPass;
 
 final class StatieKernel extends Kernel
